backend/requirements.txt
django==2.2.1
pytest==4.4.1
pytest-django==3.4.8
pytest-cov==2.6.1
mixer==6.1.3
graphene-django==2.2.0
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
django-cors-headers==2.5.3